feat: add setup command and simplify init
