+++ crates/stellar-scaffold-cli/src/commands/mod.rs
@@ -15,6 +15,7 @@ pub mod clean;
 pub mod ext;
 pub mod generate;
 pub mod init;
+pub mod setup;
 pub mod update_env;
 pub mod upgrade;
 pub mod version;
@@ -53,6 +54,9 @@ impl Root {
     pub async fn run(&mut self) -> Result<(), Error> {
         match &mut self.cmd {
             Cmd::Init(init_info) => init_info.run(&self.global_args).await?,
+            Cmd::Setup(setup_info) => {
+                setup_info.run(&self.global_args).await?;
+            }
             Cmd::Version(version_info) => version_info.run(),
             Cmd::Build(build_info) => build_info.run(&self.global_args).await?,
             Cmd::Generate(generate) => match &mut generate.cmd {
@@ -82,6 +86,8 @@ impl FromStr for Root {
 pub enum Cmd {
     /// Initialize the project
     Init(init::Cmd),
+    /// Set up an existing project (copy .env, install extensions, build contracts, git init)
+    Setup(setup::Cmd),
     /// Version of the scaffold-stellar-cli
     Version(version::Cmd),
 
@@ -112,6 +118,8 @@ pub enum Error {
     // TODO: stop using Debug for displaying errors
     #[error(transparent)]
     Init(#[from] init::Error),
+    #[error(transparent)]
+    Setup(#[from] setup::Error),
     #[error(transparent)]
     BuildContracts(#[from] build::Error),
     #[error(transparent)]
